retrieving userName and logout endpoint added
