+++ index.js
@@ -1,11 +1,12 @@
 import express from 'express'
 import { PORT } from './config/config.js'
 import morgan from 'morgan'
-import { router as usersAuthRoutes } from './routes/users/users-auth.js'
+import { router as usersAuthRoutes } from './routes/users/users-auth-router.js'
 
 const app = express()
 
 app.use(morgan('dev'))
+
 app.use('/api/auth', usersAuthRoutes)
 
 app.get('/', (req, res) => {
